1.使用QUERY BUILDER取代直接寫DQL 2.加入REPLY功能 3.修改findall()用輸入後檢查內容取代

+++ delete.php
@@ -3,28 +3,30 @@
 require_once 'bootstrap.php';
 
 if (isset($_POST['id'])) {
-    $toDelete = $entityManager->find('Message',$_POST['id']);
-    $entityManager->remove($toDelete);
-    $entityManager->flush();
+    $postId = $_POST['id'];
+    if (!is_numeric($getId)) {
+        throw new Exception('Id is not a number!');
+    }
+
+    $toDelete = $entityManager->find('Message', $postId);
+    $errMsg = NULL;
+
+    if (!$toDelete) {
+        echo 'This Id is invalid<br />';
+        $errMsg = 'error';
+    }
+
+    if (!$errMsg) {
+        $entityManager->remove($toDelete);
+        $entityManager->flush();
+        echo 'success';
+    }
 }
 ?>
-<form name="form" action=<?php echo $_SERVER['PHP_SELF'];?> method="post">
-    Select ID to delete:
-<select name="id">
-
-<?php
-$messageRep = $entityManager->getRepository('Message');
-$messages = $messageRep->findAll();
-foreach ($messages as $message) {
-    printf(
-        '<option value="%d">%d</option>',
-        $message->getId(),
-        $message->getId()
-    );
-}
-?>
-
-</select>
+<form name="form" action='delete.php' method="post">
+    Input ID to delete:
+<br />
+<input type="text" name="id"/>
 <input type="submit" value="submit" />
 </form>
 <a href="main.php">main</a>
